+++ apps/docs/app/docs/layout.tsx
@@ -20,11 +20,11 @@ const docsOptions: DocsLayoutProps = {
           ...option,
           icon: (
             <div
-              className="rounded-md p-1 shadow-lg ring-2 [&_svg]:size-6.5 md:[&_svg]:size-5"
+              className="rounded-lg p-1.5 shadow-lg ring-2 m-px border [&_svg]:size-6.5 md:[&_svg]:size-5"
               style={
                 {
                   color,
-                  border: `1px solid color-mix(in oklab, ${color} 50%, transparent)`,
+                  borderColor: `color-mix(in oklab, ${color} 50%, transparent)`,
                   '--tw-ring-color': `color-mix(in oklab, ${color} 20%, transparent)`,
                 } as object
               }