+++ packages/content-collections/src/configuration.ts
@@ -8,18 +8,9 @@ import {
   compileMDX as baseCompileMDX,
   type Options as MDXOptions,
 } from '@content-collections/mdx';
-import {
-  rehypeCode,
-  remarkGfm,
-  remarkHeading,
-  remarkImage,
-  remarkStructure,
-  type RemarkHeadingOptions,
-  type RehypeCodeOptions,
-  type StructuredData,
-  type RemarkImageOptions,
-} from 'fumadocs-core/mdx-plugins';
-import type { z as Zod } from 'zod';
+import type { StructuredData } from 'fumadocs-core/mdx-plugins';
+import * as Plugins from 'fumadocs-core/mdx-plugins';
+import { z, z as Zod } from 'zod';
 import {
   resolvePlugin,
   resolvePlugins,
@@ -35,12 +26,15 @@ export interface TransformOptions
    * Generate `structuredData`
    *
    * @defaultValue true
+   * @deprecated use `remarkStructureOptions` instead
    */
   generateStructuredData?: boolean;
 
-  remarkHeadingOptions?: RemarkHeadingOptions | boolean;
-  rehypeCodeOptions?: RehypeCodeOptions | boolean;
-  remarkImageOptions?: RemarkImageOptions | boolean;
+  remarkStructureOptions?: Plugins.StructureOptions | boolean;
+  remarkHeadingOptions?: Plugins.RemarkHeadingOptions | boolean;
+  rehypeCodeOptions?: Plugins.RehypeCodeOptions | boolean;
+  remarkImageOptions?: Plugins.RemarkImageOptions | boolean;
+  remarkCodeTabOptions?: Plugins.RemarkCodeTabOptions | boolean;
 }
 
 /**
@@ -84,10 +78,12 @@ export async function transformMDX<D extends BaseDoc>(
   }
 > {
   const {
-    generateStructuredData = true,
-    rehypeCodeOptions,
-    remarkHeadingOptions,
-    remarkImageOptions,
+    generateStructuredData,
+    remarkStructureOptions = generateStructuredData ?? true,
+    rehypeCodeOptions = true,
+    remarkHeadingOptions = true,
+    remarkImageOptions = true,
+    remarkCodeTabOptions = true,
     ...rest
   } = options;
 
@@ -111,20 +107,22 @@ export async function transformMDX<D extends BaseDoc>(
           ...rest,
           rehypePlugins: resolvePlugins(
             (plugins) => [
-              resolvePlugin(rehypeCode, rehypeCodeOptions ?? true),
+              resolvePlugin(Plugins.rehypeCode, rehypeCodeOptions),
               ...plugins,
             ],
             rest.rehypePlugins,
           ),
           remarkPlugins: resolvePlugins(
             (plugins) => [
-              remarkGfm,
-              resolvePlugin(remarkHeading, remarkHeadingOptions ?? true),
-              resolvePlugin(remarkImage, remarkImageOptions, {
+              Plugins.remarkGfm,
+              resolvePlugin(Plugins.remarkHeading, remarkHeadingOptions),
+              resolvePlugin(Plugins.remarkImage, remarkImageOptions, {
                 useImport: false,
               }),
+              'remarkCodeTab' in Plugins &&
+                resolvePlugin(Plugins.remarkCodeTab, remarkCodeTabOptions),
               ...plugins,
-              generateStructuredData && remarkStructure,
+              resolvePlugin(Plugins.remarkStructure, remarkStructureOptions),
               () => {
                 return (_, file) => {
                   data = file.data;
@@ -145,6 +143,25 @@ export async function transformMDX<D extends BaseDoc>(
     },
   );
 }
+
+export const metaSchema = z.object({
+  title: z.string().optional(),
+  pages: z.array(z.string()).optional(),
+  description: z.string().optional(),
+  root: z.boolean().optional(),
+  defaultOpen: z.boolean().optional(),
+  icon: z.string().optional(),
+});
+
+export const frontmatterSchema = z.object({
+  title: z.string(),
+  description: z.string().optional(),
+  icon: z.string().optional(),
+  full: z.boolean().optional(),
+
+  // Fumadocs OpenAPI generated
+  _openapi: z.object({}).passthrough().optional(),
+});
 
 export function createDocSchema(z: typeof Zod) {
   return {